[#IOPLT-958] Refactor routes, small app.ts
